move notes-related components into notes subfolder
