+++ client/src/components/notes/Notes/Notes.style.tsx
